+++ src/main/java/shop/wannab/book_service/book/controller/BookController.java
@@ -10,11 +10,13 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import shop.wannab.book_service.book.controller.response.BookLikeListResponse;
 import shop.wannab.book_service.book.controller.response.BookListResponse;
+import shop.wannab.book_service.book.controller.response.HotBooksListResponse;
 import shop.wannab.book_service.book.dto.BookIdListDto;
 import shop.wannab.book_service.book.dto.BookIdTitlePriceListDto;
 import shop.wannab.book_service.book.dto.OrderBookInfoListDto;
 import shop.wannab.book_service.book.dto.OrderItemListDto;
 import shop.wannab.book_service.book.controller.response.BookDetailResponse;
+import shop.wannab.book_service.book.service.impl.BookLikeServiceImpl;
 import shop.wannab.book_service.book.service.impl.BookServiceImpl;
 import shop.wannab.book_service.global.response.ApiResponse;
 
@@ -25,6 +27,7 @@ import shop.wannab.book_service.global.response.ApiResponse;
 public class BookController {
 
     protected final BookServiceImpl bookServiceImpl;
+    private final BookLikeServiceImpl bookLikeService;
 
     @PostMapping("/validation/primary")
     public void validateOrderItems(@RequestBody OrderItemListDto orderItemListDto) {
@@ -63,7 +66,7 @@ public class BookController {
     @PostMapping("/{book-id}/likes")
     public ResponseEntity<ApiResponse<Void>> createBookLike(@PathVariable("book-id") Long bookId,
                                                             @RequestHeader("X-USER-ID") Long userId) {
-        bookServiceImpl.createBookLike(bookId, userId);
+        bookLikeService.createBookLike(bookId, userId);
         return ResponseEntity.ok(ApiResponse.success(null));
     }
 
@@ -71,7 +74,7 @@ public class BookController {
     @DeleteMapping("/{book-id}/likes")
     public ResponseEntity<ApiResponse<Void>> deleteBookLike(@PathVariable("book-id") Long bookId,
                                                             @RequestHeader("X-USER-ID") Long userId) {
-        bookServiceImpl.deleteBookLike(bookId, userId);
+        bookLikeService.deleteBookLike(bookId, userId);
         return ResponseEntity.ok(ApiResponse.success(null));
     }
 
@@ -79,7 +82,7 @@ public class BookController {
     @GetMapping("/{book-id}/likes")
     public ResponseEntity<ApiResponse<Boolean>> isBookLiked(@PathVariable("book-id") Long bookId,
                                                             @RequestHeader("X-USER-ID") Long userId) {
-        Boolean isLiked = bookServiceImpl.isBookLiked(bookId, userId);
+        Boolean isLiked = bookLikeService.isBookLiked(bookId, userId);
         return ResponseEntity.ok(ApiResponse.success(isLiked));
     }
 
@@ -97,13 +100,21 @@ public class BookController {
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "10") int size) {
 
-        Page<BookLikeListResponse> bookLikes = bookServiceImpl.getLikedBooks(userId, page, size);
+        Page<BookLikeListResponse> bookLikes = bookLikeService.getLikedBooks(userId, page, size);
         return ResponseEntity.ok(ApiResponse.success(bookLikes));
     }
   
     @PostMapping("/names")
     public ResponseEntity<Map<Long, String>> getBookNames(@RequestBody List<Long> bookIds) {
         return ResponseEntity.ok(bookServiceImpl.findBookNamesByIds(bookIds));
+    }
 
+    @GetMapping("/hot-books")
+    public ResponseEntity<ApiResponse<List<HotBooksListResponse>>> getHotBooks(
+            @RequestParam(defaultValue = "0") int page,
+            @RequestParam(defaultValue = "10") int size){
+
+        List<HotBooksListResponse> hotBooksListResponse = bookLikeService.getHotBooksList(page,size);
+        return ResponseEntity.ok(ApiResponse.success(hotBooksListResponse));
     }
 }
