+++ src/main/java/shop/wannab/book_service/book/repository/BookLikeRepository.java
@@ -8,6 +8,8 @@ import org.springframework.data.repository.query.Param;
 import shop.wannab.book_service.book.entity.Book;
 import shop.wannab.book_service.book.entity.BookLike;
 
+import java.util.List;
+
 public interface BookLikeRepository extends JpaRepository<BookLike,Long> {
     // 도서 좋아요 여부 조회
      boolean existsByUserIdAndBook_BookId(Long userId, Long bookId);
@@ -18,4 +20,12 @@ public interface BookLikeRepository extends JpaRepository<BookLike,Long> {
     @Query("SELECT bl.book FROM book_like bl WHERE bl.userId = :userId")
     Page<Book> findBooksLikedByUserId(@Param("userId") Long userId, Pageable pageable);
 
+
+    @Query("""
+    SELECT bl.book
+    FROM book_like bl
+    GROUP BY bl.book
+    ORDER BY COUNT(bl) DESC
+    """)
+    List<Book> findTop10BooksByLikeCount(Pageable pageable);
 }
